+++ src/Generator.ts
@@ -3,8 +3,9 @@ import fs from 'fs-extra'
 import JSON5 from 'json5'
 import path from 'path'
 import request from 'request-promise-native'
-import {castArray, dedent, isArray, isEmpty, isFunction} from 'vtils'
+import {castArray, dedent, isArray, isEmpty, isFunction, omit} from 'vtils'
 import {CategoryList, Config, ExtendedInterface, Interface, InterfaceList, Method, PropDefinition, RequestBodyType, RequestFormItemType, Required, ResponseBodyType, ServerConfig, SyntheticalConfig} from './types'
+import {formatDate} from '@vtils/date'
 import {getNormalizedRelativePath, jsonSchemaStringToJsonSchema, jsonSchemaToType, jsonToJsonSchema, mockjsTemplateToJsonSchema, propDefinitionsToJsonSchema, throwError} from './utils'
 import {JSONSchema4} from 'json-schema'
 
@@ -204,6 +205,7 @@ export class Generator {
           dedent`
             /* tslint:disable */
             /* eslint-disable */
+            /* prettier-ignore */
 
             /* 该文件由 yapi-to-typescript 自动生成，请勿直接修改！！！ */
 
@@ -392,6 +394,13 @@ export class Generator {
       ),
     )
 
+    if (category) {
+      category.list.forEach(interfaceInfo => {
+        // 实现 _category 字段
+        interfaceInfo._category = omit(category, ['list'])
+      })
+    }
+
     return category ? category.list : []
   }
 
@@ -496,27 +505,67 @@ export class Generator {
     // 转义标题中的 /
     const escapedTitle = String(interfaceInfo.title).replace(/\//g, '\\/')
 
+    // 接口标题
+    const interfaceTitle: string = `[${escapedTitle}↗](${syntheticalConfig.serverUrl}/project/${interfaceInfo.project_id}/interface/api/${interfaceInfo._id})`
+
+    // 接口摘要
+    const interfaceSummary: Array<{
+      label: string,
+      value: string | string[],
+    }> = [
+      {
+        label: '分类',
+        value: `[${interfaceInfo._category.name}↗](${syntheticalConfig.serverUrl}/project/${interfaceInfo.project_id}/interface/api/cat_${interfaceInfo.catid})`,
+      },
+      {
+        label: '标签',
+        value: interfaceInfo.tag.map(tag => `\`${tag}\``),
+      },
+      {
+        label: '请求头',
+        value: `\`${interfaceInfo.method.toUpperCase()} ${interfaceInfo.path}\``,
+      },
+      {
+        label: '更新时间',
+        value: process.env.JEST_WORKER_ID // 测试时使用 unix 时间戳
+          ? String(interfaceInfo.up_time)
+          : `\`${formatDate(interfaceInfo.up_time, 'YYYY-MM-DD HH:mm:ss')}\``,
+      },
+    ]
+    const interfaceExtraComments: string = interfaceSummary
+      .filter(item => !isEmpty(item.value))
+      .map(item => `* @${item.label} ${castArray(item.value).join(', ')}`)
+      .join('\n')
+
     return dedent`
       /**
-       * 接口 **${escapedTitle}** 的 **请求类型**
+       * 接口 ${interfaceTitle} 的 **请求类型**
+       *
+       ${interfaceExtraComments}
        */
       ${requestDataType.trim()}
 
       /**
-       * 接口 **${escapedTitle}** 的 **返回类型**
+       * 接口 ${interfaceTitle} 的 **返回类型**
+       *
+       ${interfaceExtraComments}
        */
       ${responseDataType.trim()}
 
       ${syntheticalConfig.typesOnly ? '' : dedent`
         /**
-         * 接口 **${escapedTitle}** 的 **请求函数**
+         * 接口 ${interfaceTitle} 的 **请求函数**
+         *
+         ${interfaceExtraComments}
          */
         export function ${requestFunctionName}(requestData${isRequestDataOptional ? '?' : ''}: ${requestDataTypeName}): Promise<${responseDataTypeName}> {
           return request(prepare(${requestFunctionName}.requestConfig, requestData))
         }
 
         /**
-         * 接口 **${escapedTitle}** 的 **请求配置**
+         * 接口 ${interfaceTitle} 的 **请求配置**
+         *
+         ${interfaceExtraComments}
          */
         ${requestFunctionName}.requestConfig = Object.freeze({
           mockUrl: mockUrl${categoryUID},
@@ -539,7 +588,9 @@ export class Generator {
 
         ${(!syntheticalConfig.reactHooks || !syntheticalConfig.reactHooks.enable) ? '' : dedent`
           /**
-           * 接口 **${escapedTitle}** 的 **自动触发 API 的 React Hook**
+           * 接口 ${interfaceTitle} 的 **自动触发 API 的 React Hook**
+           *
+           ${interfaceExtraComments}
            */
           export const ${autoApiHookName} = createApiHook<typeof ${requestFunctionName}, ${isRequestDataOptional}>({
             useState: useState,
@@ -549,7 +600,9 @@ export class Generator {
           })
 
           /**
-           * 接口 **${escapedTitle}** 的 **手动触发 API 的 React Hook**
+           * 接口 ${interfaceTitle} 的 **手动触发 API 的 React Hook**
+           *
+           ${interfaceExtraComments}
            */
           export const ${manualApiHookName} = createApiHook<typeof ${requestFunctionName}, ${isRequestDataOptional}>({
             useState: useState,