+++ CHANGELOG.md
@@ -2,6 +2,52 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="1.19.2"></a>
+## [1.19.2](https://github.com/fjc0k/yapi-to-typescript/compare/v1.19.1...v1.19.2) (2019-12-12)
+
+
+
+<a name="1.19.1"></a>
+## [1.19.1](https://github.com/fjc0k/yapi-to-typescript/compare/v1.19.0...v1.19.1) (2019-12-10)
+
+
+### Bug Fixes
+
+* **ci:** 测试时使用 unix 时间戳 ([5d1edd0](https://github.com/fjc0k/yapi-to-typescript/commit/5d1edd0))
+
+
+
+<a name="1.19.0"></a>
+# [1.19.0](https://github.com/fjc0k/yapi-to-typescript/compare/v1.18.3...v1.19.0) (2019-12-10)
+
+
+### Features
+
+* 为生成的代码增加接口摘要信息 ([d6daad2](https://github.com/fjc0k/yapi-to-typescript/commit/d6daad2))
+* 生成的文件添加 prettier-ignore ([66d5156](https://github.com/fjc0k/yapi-to-typescript/commit/66d5156))
+
+
+
+<a name="1.18.3"></a>
+## [1.18.3](https://github.com/fjc0k/yapi-to-typescript/compare/v1.18.2...v1.18.3) (2019-12-04)
+
+
+### Bug Fixes
+
+* 防止 typeName 被 JSTT 转换 (close: [#17](https://github.com/fjc0k/yapi-to-typescript/issues/17)) ([58a77ac](https://github.com/fjc0k/yapi-to-typescript/commit/58a77ac))
+
+
+
+<a name="1.18.2"></a>
+## [1.18.2](https://github.com/fjc0k/yapi-to-typescript/compare/v1.18.1...v1.18.2) (2019-11-26)
+
+
+### Bug Fixes
+
+* export ApiHook ([c8a8d9a](https://github.com/fjc0k/yapi-to-typescript/commit/c8a8d9a))
+
+
+
 <a name="1.18.1"></a>
 ## [1.18.1](https://github.com/fjc0k/yapi-to-typescript/compare/v1.18.0...v1.18.1) (2019-09-17)